Imagenes en firebase

+++ app/src/main/java/com/techfind/myapplication/local/User.kt
@@ -12,7 +12,7 @@ data class User(
     @ColumnInfo(name = "name") var name: String = "",
     @ColumnInfo(name = "email") var email: String = "",
     @ColumnInfo(name = "password") var password: String = "",
-    @ColumnInfo(name = "document") var document: Int = 0,
-    @ColumnInfo(name = "cel_number") var cel_number: Int = 0,
+    @ColumnInfo(name = "document") var document: Long = 0,
+    @ColumnInfo(name = "cel_number") var cel_number: Long = 0,
 
-) : Serializable
+    ) : Serializable
